+++ tests/utils/elliptic.random.test.ts
@@ -1,28 +1,23 @@
 import { describe, expect, it } from "vitest";
 
-import { ECIES_CONFIG } from "../../src";
+import type { EllipticCurve } from "../../src/config";
 import { getValidSecret, isValidPrivateKey } from "../../src/utils";
 
 describe("test random elliptic", () => {
-  function testRandom() {
-    const key = getValidSecret();
-    expect(isValidPrivateKey(key)).toBe(true);
+  function testRandom(curve?: EllipticCurve) {
+    const key = getValidSecret(curve);
+    expect(isValidPrivateKey(key, curve)).toBe(true);
   }
 
   it("tests secp256k1", () => {
-    ECIES_CONFIG.ellipticCurve = "secp256k1";
-    testRandom();
+    testRandom("secp256k1");
   });
 
   it("tests x25519", () => {
-    ECIES_CONFIG.ellipticCurve = "x25519";
-    testRandom();
-    ECIES_CONFIG.ellipticCurve = "secp256k1";
+    testRandom("x25519");
   });
 
   it("tests ed25519", () => {
-    ECIES_CONFIG.ellipticCurve = "ed25519";
-    testRandom();
-    ECIES_CONFIG.ellipticCurve = "secp256k1";
+    testRandom("ed25519");
   });
 });